+++ rollup.browser.config.js
@@ -2,12 +2,12 @@ const minify = require('rollup-plugin-babel-minify');
 const commonjs = require('rollup-plugin-commonjs');
 
 export default {
-  input: 'lib/index.js',
+  input: 'lib/browser.js',
   output: {
-    file: 'lib/react-apollo.umd.js',
+    file: 'lib/react-apollo.browser.umd.js',
     format: 'umd',
     name: 'react-apollo',
-    sourcemap: false,
+    sourcemap: true,
   },
   plugins: [commonjs(), minify()],
   onwarn,