+++ src/renderToStringWithData.ts
@@ -1,11 +1,12 @@
-import * as React from 'react';
+import { ReactElement } from 'react';
 import * as ReactDOM from 'react-dom/server';
-import getDataFromTree from './getDataFromTree';
 
-export default function renderToStringWithData(
-  component: React.ReactElement<any>,
+import { default as getDataFromTree } from './getDataFromTree';
+
+export function renderToStringWithData(
+  component: ReactElement<any>
 ): Promise<string> {
   return getDataFromTree(component).then(() =>
-    ReactDOM.renderToString(component),
+    ReactDOM.renderToString(component)
   );
 }